+++ src/web/containers/BotDetails/BotDetails.tsx
@@ -6,8 +6,8 @@ import {
     JaiaContextType,
     JaiaDispatchContext,
     JaiaAction,
-} from "../../context/Jaia/JaiaContext";
-import { JaiaActions } from "../../context/Jaia/jaia-actions";
+} from "../../context/JaiaContext";
+import { JaiaActions } from "../../context/jaia-actions";
 import { NodeTypes } from "../../types/jaia-system-types";
 import { DETAILS_DECIMALS, UNASSIGNED_ID } from "../../utils/constants";
 import { BotAccordionNames } from "../../types/context-types";
