+++ src/web/containers/MissionsPanel/MissionsList/MissionsList.tsx
@@ -2,8 +2,8 @@
 import { useContext } from "react";
 
 // Jaia
-import { JaiaContext, JaiaDispatchContext } from "../../../context/Jaia/JaiaContext";
-import { JaiaActions } from "../../../context/Jaia/jaia-actions";
+import { JaiaContext, JaiaDispatchContext } from "../../../context/JaiaContext";
+import { JaiaActions } from "../../../context/jaia-actions";
 import MissionAssignMenu from "../../../components/MissionAssignMenu/MissionAssignMenu";
 
 import { missionsManager } from "../../../data/missions_manager/missions-manager";
