+++ src/web/containers/NodeList/__tests__/NodeList.test.tsx
@@ -2,7 +2,7 @@ import { render, screen, within } from "@testing-library/react";
 import { userEvent } from "@testing-library/user-event";
 
 import NodeList from "../NodeList";
-import { JaiaContextProvider } from "../../../context/Jaia/JaiaContext";
+import { JaiaContextProvider } from "../../../context/JaiaContext";
 
 import { PortalBotStatus } from "../../../shared/PortalStatus";
 import { PortalHubStatus } from "../../../shared/PortalStatus";
